moved requests out of utils
