+++ packages/views/src/utils/index.ts
@@ -1,5 +1,4 @@
 export * as UserAgent from './user-agent';
 export * from './user-agent';
-export * from './requests';
 export * from './currency';
 export * from './delay';
