--- packages/views/src/utils/requests/index.ts
@@ -1,5 +1,0 @@
-export * from './types';
-export * from './base';
-export * from './payment';
-export * from './login';
-export * from './tip';
